+++ pre_commit/color.py
@@ -1,3 +1,4 @@
+import argparse
 import os
 import sys
 
@@ -94,4 +95,13 @@ def use_color(setting: str) -> bool:
             terminal_supports_color and
             os.getenv('TERM') != 'dumb'
         )
+    )
+
+
+def add_color_option(parser: argparse.ArgumentParser) -> None:
+    parser.add_argument(
+        '--color', default=os.environ.get('PRE_COMMIT_COLOR', 'auto'),
+        type=use_color,
+        metavar='{' + ','.join(COLOR_CHOICES) + '}',
+        help='Whether to use color in output.  Defaults to `%(default)s`.',
     )